P-Peaks description
This is where I find out that every random homogeneous should have
been run several times, because it's bound to reach different
results. Maybe not importante in this, but it can be if we choose the
best among them, it might not be the best after all.

Besides, it would be interesting to also try and tune *by hand* just
to have an idea of where the best results are; also try to modify a
bit these best results in which we could call a Best Heterogeneous
strategy. @mariosky maybe for next EvoStar?